requirements.txt
Django==3.2.10
djangorestframework==3.13.1
numpy==1.22.3
pandas==1.4.2
requests==2.27.1
